Create `NodeSerializers` namespace
